summarization_gradio_interface_files/requirements.txt
gradio
transformers
newspaper3k
torch